@@ -1,18 +1,19 @@
 package com.edgar.filemanager;
 
 import android.os.Bundle;
-import android.support.annotation.LayoutRes;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import androidx.annotation.LayoutRes;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+
 public abstract class FileBaseFragment extends Fragment {
 
-
-    protected @LayoutRes abstract int getLayoutId();
+    protected @LayoutRes
+    abstract int getLayoutId();
 
     @Nullable
     @Override
