@@ -1,4 +1,12 @@
 package com.edgar.filemanager.classify;
 
-public class ClassifyFragment {
+import com.edgar.filemanager.FileBaseFragment;
+import com.edgar.filemanager.R;
+
+public class ClassifyFragment extends FileBaseFragment {
+
+    @Override
+    protected int getLayoutId() {
+        return R.layout.classify_fragment_layout;
+    }
 }
